Reworked Unicode; changed version format; more Py_ssize_t
Reworked Unicode support, properly differentiating between SQLWCHAR on the different platforms.
This should fix a lot of the OS/X problems and perhaps Linux UCS4 build problems.

Version format now includes the branch name if not 'master' or v<version>, allowing custom
builds to be identified, such as v2unicode-2.1.8-beta03.

Also tested with 64-bit Windows builds, so more Py_ssize_t warnings were found and corrected.

Created TRACE macro to replace the #ifdefs sprinkled through the code.

+++ src/row.cpp
@@ -39,7 +39,7 @@ struct Row
 #define Row_CheckExact(op) ((op)->ob_type == &RowType)
 
 void
-FreeRowValues(int cValues, PyObject** apValues)
+FreeRowValues(Py_ssize_t cValues, PyObject** apValues)
 {
     if (apValues)
     {
